feat(ai-tier): PR1b — relabel 12 genuine pack long-answers Short->Long; quarantine 7 (#253)
AI-tier PR1b (pack-file relabel-only). Authority: report-aitier-pr1-mechanical-2026-06-17.md
(the 19-item PACK_5MK_SHORT_BACKLOG) + cofounder Group-A/B classification.

Group A — relabel format:"Short" -> "Long" (12 genuine 5-mark long-answers; label-only,
no content/marks/count change). Confirmed each by reading its questionText:
  ARC2-016, ARC2-017, ABS2-048, CC2-048, CR2-044, CR2-045, CR2-046,
  HEC2-039, LT2-016, LT2-024, ME2-025, REP2-048

Reclassified on inspection: PR2-018 ("3 red, 4 green, 5 blue: P(not blue)") is a
single-step 7/12 one-liner, NOT a 5-mark long-answer -> MOVED to Group B (quarantine),
not relabelled, per the instruction's safeguard.

Group B — quarantine (content<->marks mismatch; NOT relabelled). Kept in
PACK_5MK_SHORT_BACKLOG with an annotation; logged as [FU-AITIER-MARKS-MISMATCH]:
  TG3-056, TG3-059, ABS2-047, CR2-043, MNM2-037, REP2-039, PR2-018

Net: backlog 19 -> 7. Count unchanged (label-only). Guard still PASS (set == backlog).

Gates: tsc PASS; root matrix 181/181; ops matrix PASS; mojibake PASS;
scope:guard mixed PASS; diff --check clean. No forbidden files touched.

+++ lazytopper/src/data/questionBanks/class10/science/acidsBasesSalts.pack2.ts
@@ -1409,7 +1409,7 @@ export const ABS2_PACK2: CanonicalQuestion[] = [
     "subtopic": "Properties of Acids and Bases",
     "section": "D",
     "marks": 5,
-    "format": "Short",
+    "format": "Long",
     "difficulty": "Medium",
     "bloomSkill": "Applying",
     "questionText": "A student accidentally spills concentrated sulfuric acid on the laboratory bench. Explain the steps to safely neutralize the acid using a commonly available base.",
